+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+- add keys and locked doors
+- add tram pass
+- add warp from teacher's lounge to cpu wing
+
 ## [0.0.10] - 2024-01-16
 
 - make enemies individually scalable